Enable running tests without an allocator
Currently we always run tests with "std" enabled. It is currently not
possible to run the tests without an allocator but there is no reason
that we should not be able to run allocation free tests without an
allocator, doing so is better because it tests our claim that the lib
can be used without an allocator.