Add all_zeros method to Hash trait
Currently we use `Default::default()` to get an all zeros hash. This is
misleading because an all zeroes hash does not have a real meaning since
there is no known input to create it. The all zeros has is used by the
Bitcoin network to signify certain things (e.g. the previous blockhash
in the genesis block).

Remove the `Default` implementation from the `Hash` trait and add a
`all_zeros` method instead.