Deduplicate hash types
There was a lot of code duplicated for all hash types. Most of it
various standard impls. This made it more tedious and error-prone to
modify the impls for all types.

This commit creates two new macros that take care of implementing the
hash types and calls them from each hash module as needed to keep API
and behavior without a change.

This also extends the `serde_impl!` macro to support generics.